Move job runners into common sub directory
