+++ qt/scientific_interfaces/ISISReflectometry/GUI/Batch/RowPreprocessingAlgorithm.h
@@ -8,7 +8,7 @@
 
 #include "Common/DllConfig.h"
 #include "MantidAPI/IAlgorithm_fwd.h"
-#include "MantidQtWidgets/Common/BatchAlgorithmRunner.h"
+#include "MantidQtWidgets/Common/JobRunners/BatchAlgorithmRunner.h"
 
 #include <string>
 #include <vector>
