+++ qt/scientific_interfaces/ISISReflectometry/GUI/Batch/SumBanksAlgorithm.h
@@ -9,7 +9,7 @@
 
 #include "Common/DllConfig.h"
 #include "MantidAPI/IAlgorithm_fwd.h"
-#include "MantidQtWidgets/Common/BatchAlgorithmRunner.h"
+#include "MantidQtWidgets/Common/JobRunners/BatchAlgorithmRunner.h"
 
 #include <string>
 #include <vector>
