+++ qt/scientific_interfaces/ISISReflectometry/GUI/Runs/QtCatalogSearcher.cpp
@@ -11,8 +11,8 @@
 #include "MantidAPI/CatalogManager.h"
 #include "MantidAPI/ITableWorkspace.h"
 #include "MantidQtWidgets/Common/AlgorithmDialog.h"
-#include "MantidQtWidgets/Common/AlgorithmRunner.h"
 #include "MantidQtWidgets/Common/InterfaceManager.h"
+#include "MantidQtWidgets/Common/JobRunners/AlgorithmRunner.h"
 
 #include <boost/regex.hpp>
