+++ qt/scientific_interfaces/ISISReflectometry/GUI/Runs/QtRunsView.cpp
@@ -7,10 +7,10 @@
 #include "QtRunsView.h"
 #include "MantidKernel/UsageService.h"
 #include "MantidQtIcons/Icon.h"
-#include "MantidQtWidgets/Common/AlgorithmRunner.h"
 #include "MantidQtWidgets/Common/FileDialogHandler.h"
 #include "MantidQtWidgets/Common/HelpWindow.h"
 #include "MantidQtWidgets/Common/HintingLineEditFactory.h"
+#include "MantidQtWidgets/Common/JobRunners/AlgorithmRunner.h"
 #include <QMenu>
 #include <QMessageBox>
