+++ qt/scientific_interfaces/Indirect/InelasticDataManipulationSqwTabModel.h
@@ -9,7 +9,7 @@
 #include "DllConfig.h"
 #include "IndirectDataValidationHelper.h"
 #include "MantidGeometry/IComponent.h"
-#include "MantidQtWidgets/Common/BatchAlgorithmRunner.h"
+#include "MantidQtWidgets/Common/JobRunners/BatchAlgorithmRunner.h"
 #include "MantidQtWidgets/Common/QtPropertyBrowser/QtTreePropertyBrowser"
 #include <typeinfo>
