+++ qt/scientific_interfaces/Indirect/InelasticDataManipulationSymmetriseTabModel.h
@@ -8,7 +8,7 @@
 
 #include "DllConfig.h"
 #include "IndirectDataValidationHelper.h"
-#include "MantidQtWidgets/Common/BatchAlgorithmRunner.h"
+#include "MantidQtWidgets/Common/JobRunners/BatchAlgorithmRunner.h"
 #include "MantidQtWidgets/Common/QtPropertyBrowser/QtTreePropertyBrowser"
 #include <typeinfo>
