+++ qt/widgets/common/inc/MantidQtWidgets/Common/JobRunners/BatchAlgorithmRunner.h
@@ -6,9 +6,9 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "DllOption.h"
-#include "IConfiguredAlgorithm.h"
 #include "MantidAPI/Algorithm.h"
+#include "MantidQtWidgets/Common/DllOption.h"
+#include "MantidQtWidgets/Common/IConfiguredAlgorithm.h"
 
 #include <QObject>
