+++ qt/widgets/common/inc/MantidQtWidgets/Common/JobRunners/IJobRunner.h
@@ -6,7 +6,7 @@
 // SPDX - License - Identifier: GPL - 3.0 +
 #pragma once
 
-#include "DllOption.h"
+#include "MantidQtWidgets/Common/DllOption.h"
 
 #include <deque>
 #include <memory>
