+++ qt/widgets/common/test/BatchAlgorithmRunnerTest.h
@@ -14,7 +14,7 @@
 #include "MantidAPI/MatrixWorkspace.h"
 #include "MantidAPI/WorkspaceHistory.h"
 #include "MantidQtWidgets/Common/AlgorithmRuntimeProps.h"
-#include "MantidQtWidgets/Common/BatchAlgorithmRunner.h"
+#include "MantidQtWidgets/Common/JobRunners/BatchAlgorithmRunner.h"
 #include "MockConfiguredAlgorithm.h"
 
 #include <QSignalSpy>
